Changed the implementation of handlers
