+++ internal/api/handlers/router.go
@@ -1,0 +1,13 @@
+package handler
+
+import (
+	"net/http"
+
+	"github.com/go-chi/chi/v5"
+)
+
+var webDir = "./web"
+
+func SetHandlers(router chi.Router) {
+	FileServer(router, "/", http.Dir(webDir))
+}
